Added Bill tests with some additional input validations

+++ pkg/bill/bill.go
@@ -2,17 +2,20 @@ package bill
 
 import (
 	"context"
+	"log"
 	"time"
 
 	"github.com/sanchitdeora/budget-tracker/db"
 	"github.com/sanchitdeora/budget-tracker/models"
+	"github.com/sanchitdeora/budget-tracker/pkg/exceptions"
 	"github.com/sanchitdeora/budget-tracker/pkg/transaction"
 	"github.com/sanchitdeora/budget-tracker/utils"
 )
 
+//go:generate mockgen -destination=./mocks/mock_bill.go -package=mock_bill github.com/sanchitdeora/budget-tracker/pkg/bill Service
 type Service interface {
-	GetBills(ctx context.Context, bill *[]models.Bill) (error)
-	GetBillById(ctx context.Context, id string, bill *models.Bill) (error)
+	GetBills(ctx context.Context) (*[]models.Bill, error)
+	GetBillById(ctx context.Context, id string) (*models.Bill, error)
 	CreateBillByUser(ctx context.Context, bill models.Bill) (string, error)
 	CreateBill(ctx context.Context, bill models.Bill) (string, error)
 	UpdateBillById(ctx context.Context, id string, bill models.Bill) (string, error)
@@ -35,18 +38,23 @@ func NewService(opts *Opts) Service {
 }
 
 
-func (s *serviceImpl) GetBills(ctx context.Context, bill *[]models.Bill) (error) {
-	// TODO: input validation
-	return s.DB.GetAllBillRecords(ctx, bill)
+func (s *serviceImpl) GetBills(ctx context.Context) (*[]models.Bill, error) {
+	return s.DB.GetAllBillRecords(ctx)
 }
 
-func (s *serviceImpl) GetBillById(ctx context.Context, id string, bill *models.Bill) (error) {
-	// TODO: input validation
-	return s.DB.GetBillRecordById(ctx, id, bill)
+func (s *serviceImpl) GetBillById(ctx context.Context, id string) (*models.Bill, error) {
+	if id == "" {
+		log.Println("Missing Bill Id")
+		return nil, exceptions.ErrValidationError
+	}	
+	return s.DB.GetBillRecordById(ctx, id)
 }
 
 func (s *serviceImpl) CreateBillByUser(ctx context.Context, bill models.Bill) (string, error) {
-	// TODO: input validation
+	if !bill.IsValid() {
+		return "", exceptions.ErrValidationError
+	}
+
 	bill.SetByUser()
 	bill.SetCategory()
 	bill.SetFrequency()
@@ -54,33 +62,51 @@ func (s *serviceImpl) CreateBillByUser(ctx context.Context, bill models.Bill) (s
 }
 
 func (s *serviceImpl) CreateBill(ctx context.Context, bill models.Bill) (string, error) {
-	// TODO: input validation
+	if !bill.IsValid() {
+		return "", exceptions.ErrValidationError
+	}
+
 	bill.SetCategory()
 	bill.SetFrequency()
 	return s.DB.InsertBillRecord(ctx, bill)
 }
 
 func (s *serviceImpl) UpdateBillById(ctx context.Context, id string, bill models.Bill) (string, error) {
-	// TODO: input validation
+	if id == "" {
+		log.Println("Missing Bill Id")
+		return "", exceptions.ErrValidationError
+	}
+
+	if !bill.IsValid() {
+		return "", exceptions.ErrValidationError
+	}
+	
 	bill.SetCategory()
 	bill.SetFrequency()
 	return s.DB.UpdateBillRecordById(ctx, id, bill)
 }
 
 func (s *serviceImpl) UpdateBillIsPaid(ctx context.Context, id string) (string, error) {
-	// TODO: input validation
-	datePaid := time.Now()
+	if id == "" {
+		log.Println("Missing Bill Id")
+		return "", exceptions.ErrValidationError
+	}
 
-	var bill models.Bill
-	s.DB.GetBillRecordById(ctx, id, &bill)
+	var bill *models.Bill
+	bill, err := s.DB.GetBillRecordById(ctx, id)
+	if err != nil {
+		log.Println("error while fetching bill record", err)
+		return "", err
+	}
 	
+	datePaid := time.Now()
 	if !bill.IsPaid {
-
 		var newTransaction models.Transaction
-		newTransaction.FromBill(bill, datePaid)
+		newTransaction.FromBill(*bill, datePaid)
 		_, err := s.TransactionService.CreateTransaction(ctx, newTransaction)
 		if err != nil {
-			return "", nil
+			log.Println("error while creating transaction record", err)
+			return "", err
 		}
 	}
 
@@ -88,7 +114,8 @@ func (s *serviceImpl) UpdateBillIsPaid(ctx context.Context, id string) (string, 
 		// create new bill entry for next frequency period
 		newDueDate, err := utils.CalculateEndDateWithFrequency(bill.DueDate, bill.Frequency)
 		if err != nil {
-			return "", nil
+			log.Println("error while calculating end date with frequency: ", bill.Frequency, "error: ", err)
+			return "", err
 		}
 
 		newBill := bill
@@ -99,9 +126,10 @@ func (s *serviceImpl) UpdateBillIsPaid(ctx context.Context, id string) (string, 
 		newBill.SequenceNumber = bill.SequenceNumber + 1
 		newBill.SequenceStartId = bill.SequenceStartId
 
-		_, err = s.CreateBill(ctx, newBill)
+		_, err = s.CreateBill(ctx, *newBill)
 		if err != nil {
-			return "", nil
+			log.Println("error while creating new bill record", err)
+			return "", err
 		}
 	}
 
@@ -109,11 +137,19 @@ func (s *serviceImpl) UpdateBillIsPaid(ctx context.Context, id string) (string, 
 }
 
 func (s *serviceImpl) UpdateBillIsUnpaid(ctx context.Context, id string) (string, error) {
-	// TODO: input validation
+	if id == "" {
+		log.Println("Missing Bill Id")
+		return "", exceptions.ErrValidationError
+	}
+
 	return s.DB.UpdateBillRecordIsUnpaid(ctx, id)
 }
 
 func (s *serviceImpl) DeleteBillById(ctx context.Context, id string) (string, error) {
-	// TODO: input validation
+	if id == "" {
+		log.Println("Missing Bill Id")
+		return "", exceptions.ErrValidationError
+	}
+
 	return s.DB.DeleteBillRecordById(ctx, id)
 }
